backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
openai
psycopg2-binary
python-dotenv
django-filter
gunicorn
